Keep all serialization tests in one place
Has the additional bonus of not having a name clash which pytest doesn't like
